feat(MEO-16): include simple loading screen
