--- assets/css/core.css
@@ -1,21 +1,0 @@
-body {
-    display: flex;
-    width: 100%;
-    height: 100%;
-    justify-content: center;
-    align-items: center;
-    flex-direction: column;
-    gap: 25px;
-    background: #282c32;
-}
-
-#mira-text {
-    display: none;
-    font-size: 60px;
-    color: white;
-}
-
-#bevy-logo {
-    display: block;
-    background: url("images/bevy_logo_dark.png");
-}
