+++ assets/css/loading_screen.css
@@ -1,0 +1,17 @@
+body {
+    width: 100%;
+    height: 100%;
+}
+
+.overlay {
+    width: 100%;
+    height: 100%;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    background: rgba(0, 0, 0, 1);
+}
+
+h2 {
+    font-size: 64px;
+}
