+++ assets/html/loading_screen.html
@@ -1,0 +1,10 @@
+<head>
+    <meta name="loading_screen">
+    <link rel="stylesheet" href="assets/css/loading_screen.css">
+    <title>Title</title>
+</head>
+<body>
+<div class="overlay">
+    <h2>Loading...</h2>
+</div>
+</body>
