--- assets/html/splashscreen.html
@@ -1,14 +1,0 @@
-<head>
-    <meta name="splash-screen">
-    <title>Splashscreen</title>
-</head>
-
-<body>
-<!-- Our name -->
-<h1 id="mira-text">Mirror Beast</h1>
-
-<!-- Bevy official logo -->
-<div id="bevy-logo"></div>
-
-
-</body>
