+++ crates/game_ui/src/lib.rs
@@ -4,7 +4,7 @@ mod screens;
 
 use bevy::prelude::*;
 use bevy_extended_ui::ExtendedUiPlugin;
-use crate::screens::SplashScreen;
+use crate::screens::ScreenPlugin;
 
 pub struct GameUiPlugin;
 
@@ -13,6 +13,6 @@ impl Plugin for GameUiPlugin {
     #[coverage(off)]
     fn build(&self, app: &mut App) {
         app.add_plugins(ExtendedUiPlugin);
-        app.add_plugins(SplashScreen);
+        app.add_plugins(ScreenPlugin);
     }
 }
